+++ familyhelper-life-impl/src/main/java/com/dwarfeng/familyhelper/life/impl/handler/PbNodeOperateHandlerImpl.java
@@ -17,14 +17,14 @@ public class PbNodeOperateHandlerImpl implements PbNodeOperateHandler {
 
     private final PbNodeMaintainService pbNodeMaintainService;
 
-    private final OperateHandlerValidator operateHandlerValidator;
+    private final HandlerValidator handlerValidator;
 
     public PbNodeOperateHandlerImpl(
             PbNodeMaintainService pbNodeMaintainService,
-            OperateHandlerValidator operateHandlerValidator
+            HandlerValidator handlerValidator
     ) {
         this.pbNodeMaintainService = pbNodeMaintainService;
-        this.operateHandlerValidator = operateHandlerValidator;
+        this.handlerValidator = handlerValidator;
     }
 
     @Override
@@ -34,21 +34,21 @@ public class PbNodeOperateHandlerImpl implements PbNodeOperateHandler {
             LongIdKey parentKey = pbNodeCreateInfo.getParentKey();
 
             // 1. 确认用户存在。
-            operateHandlerValidator.makeSureUserExists(userKey);
+            handlerValidator.makeSureUserExists(userKey);
 
             // 2. 确认个人最佳集合存在。
-            operateHandlerValidator.makeSurePbSetExists(setKey);
+            handlerValidator.makeSurePbSetExists(setKey);
 
             // 3. 确认父个人最佳节点存在。
             if (Objects.nonNull(parentKey)) {
-                operateHandlerValidator.makeSurePbNodeExists(parentKey);
+                handlerValidator.makeSurePbNodeExists(parentKey);
             }
 
             // 4. 确认用户有权限操作指定的个人最佳集合。
-            operateHandlerValidator.makeSureUserModifyPermittedForPbSet(userKey, setKey);
+            handlerValidator.makeSureUserModifyPermittedForPbSet(userKey, setKey);
 
             // 5. 确认个人最佳节点与父个人最佳节点的个人最佳集合存在。
-            operateHandlerValidator.makeSurePbSetIdenticalForPbSet(parentKey, setKey);
+            handlerValidator.makeSurePbSetIdenticalForPbSet(parentKey, setKey);
 
             // 6. 根据 pbNodeCreateInfo 以及创建的规则组合 个人最佳节点 实体。
             PbNode pbNode = new PbNode(
@@ -71,21 +71,21 @@ public class PbNodeOperateHandlerImpl implements PbNodeOperateHandler {
             LongIdKey parentKey = pbNodeUpdateInfo.getParentKey();
 
             // 1. 确认用户存在。
-            operateHandlerValidator.makeSureUserExists(userKey);
+            handlerValidator.makeSureUserExists(userKey);
 
             // 2. 确认个人最佳节点存在。
-            operateHandlerValidator.makeSurePbNodeExists(pbNodeKey);
+            handlerValidator.makeSurePbNodeExists(pbNodeKey);
 
             // 3. 确认父个人最佳节点存在。
             if (Objects.nonNull(parentKey)) {
-                operateHandlerValidator.makeSurePbNodeExists(parentKey);
+                handlerValidator.makeSurePbNodeExists(parentKey);
             }
 
             // 4. 确认用户有权限操作指定的个人最佳节点。
-            operateHandlerValidator.makeSureUserModifyPermittedForPbNode(userKey, pbNodeKey);
+            handlerValidator.makeSureUserModifyPermittedForPbNode(userKey, pbNodeKey);
 
             // 5. 确认个人最佳节点与父个人最佳节点的个人最佳集合存在。
-            operateHandlerValidator.makeSurePbSetIdenticalForPbNode(parentKey, pbNodeKey);
+            handlerValidator.makeSurePbSetIdenticalForPbNode(parentKey, pbNodeKey);
 
             // 6. 根据 pbNodeUpdateInfo 以及更新的规则设置 个人最佳节点 实体。
             PbNode pbNode = pbNodeMaintainService.get(pbNodeKey);
@@ -106,13 +106,13 @@ public class PbNodeOperateHandlerImpl implements PbNodeOperateHandler {
     public void removePbNode(StringIdKey userKey, LongIdKey pbNodeKey) throws HandlerException {
         try {
             // 1. 确认用户存在。
-            operateHandlerValidator.makeSureUserExists(userKey);
+            handlerValidator.makeSureUserExists(userKey);
 
             // 2. 确认个人最佳节点存在。
-            operateHandlerValidator.makeSurePbNodeExists(pbNodeKey);
+            handlerValidator.makeSurePbNodeExists(pbNodeKey);
 
             // 3. 确认用户有权限操作指定的银行卡。
-            operateHandlerValidator.makeSureUserModifyPermittedForPbNode(userKey, pbNodeKey);
+            handlerValidator.makeSureUserModifyPermittedForPbNode(userKey, pbNodeKey);
 
             // 4. 存在删除指定的个人最佳节点。
             pbNodeMaintainService.deleteIfExists(pbNodeKey);